+++ CHANGELOG.md
@@ -6,16 +6,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.0.3] - 2019-04-20
+
 ### Added
 
-- added gedmo tree for representing trees in the database
+- gedmo tree for representing trees in the database
+- Repository, Service and ServiceField entities
 
 ## [0.0.2] - 2019-04-16
 
 ### Added
 
 - jwt authentication
-- added projects and customer entity
+- projects and customer entity
 
 ## [0.0.1] - 2019-03-31
 
@@ -23,6 +26,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - base app
 
-[Unreleased]: https://github.com/siewert87/aaas-api/compare/v0.0.2..HEAD
+[Unreleased]: https://github.com/siewert87/aaas-api/compare/v0.0.3..HEAD
+[0.0.3]: https://github.com/siewert87/aaas-api/compare/v0.0.2..v0.0.3
 [0.0.2]: https://github.com/siewert87/aaas-api/compare/v0.0.1..v0.0.2
 [0.0.1]: https://github.com/siewert87/aaas-api/releases/tag/v0.0.1